requirements.txt
astroid==2.4.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1